# internal/songs/ports/http/contracts/openapi_types.gen.go
// Package contracts provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package contracts

// Pagination defines model for Pagination.
type Pagination struct {
	Limit      int `json:"limit"`
	Offset     int `json:"offset"`
	TotalCount int `json:"totalCount"`
}

// PostSongRequestBody defines model for PostSongRequestBody.
type PostSongRequestBody struct {
	// Group Song artist(-s)
	Group string `json:"group" validate:"required,min=1,max=32"`

	// Song Song title
	Song string `json:"song" validate:"required,min=1,max=32"`
}

// Song defines model for Song.
type Song struct {
	Group       string `json:"group"`
	Id          int    `json:"id"`
	Link        string `json:"link"`
	ReleaseDate string `json:"releaseDate"`
	Song        string `json:"song"`
	Text        string `json:"text"`
}

// UpdateSongRequestBody defines model for UpdateSongRequestBody.
type UpdateSongRequestBody struct {
	// Group Song artist(-s)
	Group *string `json:"group,omitempty" validate:"omitempty,min=1,max=32"`

	// Link Song link
	Link *string `json:"link,omitempty" validate:"omitempty,url"`

	// ReleaseDate Release date (DD.MM.YYYY)
	ReleaseDate *string `json:"releaseDate,omitempty" validate:"omitempty,dmyDate"`

	// Song Song title
	Song *string `json:"song,omitempty" validate:"omitempty,min=1,max=32"`

	// Text Lyrics
	Text *string `json:"text,omitempty" validate:"omitempty,max=2048"`
}

// GetSongsParams defines parameters for GetSongs.
type GetSongsParams struct {
	// Group Song group search term
	Group *string `form:"group,omitempty" json:"group,omitempty"`

	// Song Song name search term
	Song *string `form:"song,omitempty" json:"song,omitempty"`

	// Text Song lyrics search term
	Text *string `form:"text,omitempty" json:"text,omitempty"`

	// Link Song link search term
	Link *string `form:"link,omitempty" json:"link,omitempty"`

	// Before Search songs released before a specific date (DD.MM.YYYY)
	Before *string `form:"before,omitempty" json:"before,omitempty" validate:"omitempty,dmyDate"`

	// After Search songs released after a specific date (DD.MM.YYYY)
	After *string `form:"after,omitempty" json:"after,omitempty" validate:"omitempty,dmyDate"`

	// Limit Limit for pagination
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Offset for pagination
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// GetSongTextParams defines parameters for GetSongText.
type GetSongTextParams struct {
	// Limit Limit for pagination
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Offset for pagination
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateSongJSONRequestBody defines body for CreateSong for application/json ContentType.
type CreateSongJSONRequestBody = PostSongRequestBody

// UpdateSongJSONRequestBody defines body for UpdateSong for application/json ContentType.
type UpdateSongJSONRequestBody = UpdateSongRequestBody